# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TransMóvil - Transport & Local Services Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user profile
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
    patch:
      operationId: updateMe
      tags: [auth]
      summary: Update current user profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/trips:
    get:
      operationId: getMyTrips
      tags: [auth]
      summary: Get current user's trip history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"
  /users/me/favorites:
    get:
      operationId: getMyFavorites
      tags: [auth]
      summary: Get saved favorite locations
      responses:
        "200":
          description: Favorite locations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FavoriteLocation"
    post:
      operationId: addFavorite
      tags: [auth]
      summary: Save a favorite location
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/FavoriteLocationInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FavoriteLocation"
  /users/me/favorites/{id}:
    delete:
      operationId: deleteFavorite
      tags: [auth]
      summary: Remove a favorite location
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── DRIVERS ──────────────────────────────────────────────────────────────
...
  /drivers/me/trips:
    get:
      operationId: getMyDriverTrips
      tags: [drivers]
      summary: Get driver's service history
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trip list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Trip"

  # ─── VEHICLES ─────────────────────────────────────────────────────────────
  /vehicles:
    post:
      operationId: addVehicle
      tags: [vehicles]
      summary: Add a vehicle (driver)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
  /vehicles/{id}:
    get:
      operationId: getVehicle
      tags: [vehicles]
      summary: Get vehicle details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vehicle
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"
    patch:
      operationId: updateVehicle
      tags: [vehicles]
      summary: Update vehicle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VehicleUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vehicle"

  # ─── TRIPS ────────────────────────────────────────────────────────────────
  /trips:
    post:
      operationId: requestTrip
      tags: [trips]
      summary: Request a new trip (customer)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripRequest"
      responses:
        "201":
          description: Trip created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}:
    get:
      operationId: getTrip
      tags: [trips]
      summary: Get trip details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: Not found
  /trips/{id}/status:
    patch:
      operationId: updateTripStatus
      tags: [trips]
      summary: Update trip status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripStatusUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/{id}/cancel:
    patch:
      operationId: cancelTrip
      tags: [trips]
      summary: Cancel a trip
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripCancelInput"
      responses:
        "200":
          description: Cancelled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
  /trips/active:
    get:
      operationId: getActiveTrip
      tags: [trips]
      summary: Get the current user's active trip
      responses:
        "200":
          description: Active trip or null
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Trip"
        "404":
          description: No active trip

  # ─── SERVICE TYPES ────────────────────────────────────────────────────────
...
  /admin/trips:
    get:
      operationId: getAdminTrips
      tags: [admin]
      summary: List all trips
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Trips list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripListResponse"
...
components:
  schemas:
...
    UserProfile:
      type: object
      required: [id, clerkId, role, firstName, lastName, email, createdAt]
      properties:
        id:
          type: integer
        clerkId:
          type: string
        role:
          type: string
          enum: [customer, driver, admin]
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
    UserProfileUpdate:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
        avatarUrl:
          type: string
    FavoriteLocation:
      type: object
      required: [id, userId, label, address, lat, lng, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        label:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        icon:
          type: ["string", "null"]
        createdAt:
          type: string
    FavoriteLocationInput:
      type: object
      required: [label, address, lat, lng]
      properties:
        label:
          type: string
        address:
          type: string
        lat:
          type: number
        lng:
          type: number
        icon:
          type: string
    DriverPublic:
      type: object
      required: [id, firstName, lastName, rating, totalTrips, isAvailable, serviceTypeId, lat, lng]
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        totalTrips:
          type: integer
        isAvailable:
          type: boolean
        serviceTypeId:
          type: integer
        serviceTypeName:
          type: string
        vehicleMake:
          type: ["string", "null"]
        vehicleModel:
          type: ["string", "null"]
        vehicleColor:
          type: ["string", "null"]
        vehiclePlate:
          type: ["string", "null"]
        vehiclePhotoUrl:
          type: ["string", "null"]
        lat:
          type: ["number", "null"]
        lng:
          type: ["number", "null"]
...
    Vehicle:
      type: object
      required: [id, driverId, make, model, year, plate, color, serviceTypeId]
      properties:
        id:
          type: integer
        driverId:
          type: integer
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        color:
          type: string
        serviceTypeId:
          type: integer
        photoUrl:
          type: ["string", "null"]
        capacity:
          type: integer
        createdAt:
          type: string
    VehicleInput:
      type: object
      required: [make, model, year, plate, color, serviceTypeId]
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        color:
          type: string
        serviceTypeId:
          type: integer
        photoUrl:
          type: string
        capacity:
          type: integer
    VehicleUpdate:
      type: object
      properties:
        make:
          type: string
        model:
          type: string
        year:
          type: integer
        plate:
          type: string
        color:
          type: string
        photoUrl:
          type: string
        capacity:
          type: integer
    Trip:
      type: object
      required: [id, customerId, status, pickupAddress, pickupLat, pickupLng, destinationAddress, destinationLat, destinationLng, serviceTypeId, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        driverId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [searching, accepted, arriving, in_progress, completed, cancelled]
        pickupAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        serviceTypeId:
          type: integer
        estimatedPrice:
          type: ["number", "null"]
        finalPrice:
          type: ["number", "null"]
        paymentMethod:
          type: ["string", "null"]
        cancelReason:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        acceptedAt:
          type: ["string", "null"]
        startedAt:
          type: ["string", "null"]
        completedAt:
          type: ["string", "null"]
        customer:
          $ref: "#/components/schemas/UserProfile"
        driver:
          $ref: "#/components/schemas/DriverPublic"
        serviceType:
          $ref: "#/components/schemas/ServiceType"
    TripRequest:
      type: object
      required: [pickupAddress, pickupLat, pickupLng, destinationAddress, destinationLat, destinationLng, serviceTypeId]
      properties:
        pickupAddress:
          type: string
        pickupLat:
          type: number
        pickupLng:
          type: number
        destinationAddress:
          type: string
        destinationLat:
          type: number
        destinationLng:
          type: number
        serviceTypeId:
          type: integer
        paymentMethod:
          type: string
          enum: [cash, mercadopago, card]
          default: cash
        notes:
          type: string
    TripStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted, arriving, in_progress, completed]
        driverId:
          type: integer
    TripCancelInput:
      type: object
      properties:
        reason:
          type: string
    ServiceType:
      type: object
      required: [id, name, icon, description, isActive]
      properties:
        id:
          type: integer
        name:
          type: string
        icon:
          type: string
        description:
          type: string
        basePrice:
          type: number
        pricePerKm:
          type: number
        isActive:
          type: boolean
...
    TripListResponse:
      type: object
      required: [data, total]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Trip"
        total:
          type: integer
...