# lib/api-zod/src/generated/types/vehicleUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * TransMóvil - Transport & Local Services Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface VehicleUpdate {
  make?: string;
  model?: string;
  year?: number;
  plate?: string;
  color?: string;
  photoUrl?: string;
  capacity?: number;
}